Improve type signature for return value to be based on first arg
Improves the type signature of the stringify function to take into account the main argument
that we are serializing. If it is a primitive scalar, array, or object, and there is no second
argument then the return value will always be a string.

As this is the common case for user applications, this improves the user experience as we do
not have to check the result to ensure the value is defined. For most use cases where the
callsite has a primitive, array, or object and is using the default serilaization, the
return value will be known to be a string